order successful template render

+++ orders/models.py
@@ -31,7 +31,7 @@ class OrderItem(models.Model):
     order = models.ForeignKey(Order, related_name='items', on_delete=models.CASCADE)
     product = models.ForeignKey(Product, related_name='order_items', on_delete=models.CASCADE)
     price = models.DecimalField(max_digits=10, decimal_places=2)
-    quantity = models.PositiveBigIntegerField(default=1)
+    quantity = models.PositiveIntegerField(default=1)
     
     def get_cost(self):
         return self.price * self.quantity
